Fixing compiler warnings and setting warning levels - windows

+++ src/Image.cpp
@@ -14,10 +14,10 @@ void Image::create(unsigned int _width, unsigned int _height, const Colour& _col
 		
 		for (unsigned int i = 0; i < m_Pixels.size(); i += 4)
 		{
-			m_Pixels[i * 4 + 0] = _colour.r;
-			m_Pixels[i * 4 + 1] = _colour.g;
-			m_Pixels[i * 4 + 2] = _colour.b;
-			m_Pixels[i * 4 + 3] = _colour.a;
+			m_Pixels[i * 4 + 0] = static_cast<uint8_t>(_colour.r);
+			m_Pixels[i * 4 + 1] = static_cast<uint8_t>(_colour.g);
+			m_Pixels[i * 4 + 2] = static_cast<uint8_t>(_colour.b);
+			m_Pixels[i * 4 + 3] = static_cast<uint8_t>(_colour.a);
 		}
 	}
 	else
